[Nix] New private config split structure
This gets rid of the use of get-flake to evaluate the base flake,
in order to avoid excessive IFD. Now the base flake and private flake
will have same flake inputs and the flake.lock is a symlink in the private flake.
The base flake is evaluated by importing and passing inputs manually.